complete about_modules
This was pretty easy to complete but I'll have to put them into practice
to fully grok the purpose of modules. Seems like a way to include methods
in other classes without using super classes. And modules cannot be
instantiated therefore they cannot be used to create objects. In other
words, it's a way to add functionality to other, even multiple, classes.